Moved scheduling.ts
