+++ scripts/debug/eval.ts
@@ -55,7 +55,7 @@ import { onJoinData } from "../data/onjoindata.js";
 import salvageable from "../data/salvageable.js";
 import { whitelist as whitelistItems } from "../data/whitelistitems.js";
 import { xrayblocks as xrayBlocks } from "../data/xray.js";
-import * as scheduling from "../misc/scheduling.js";
+import * as scheduling from "../libs/scheduling.js";
 import scoreboard from "../libs/scoreboardnew.js";
 import * as util from "../util.js";
 import { world } from "mojang-minecraft";
