+++ scripts/penrose/tickevent/ban/serverban.ts
@@ -1,6 +1,6 @@
 import { world, EntityQueryOptions } from "mojang-minecraft";
 import { banMessage, sendMsg, sendMsgToPlayer } from "../../../util.js";
-import { setTickInterval } from "../../../misc/scheduling.js";
+import { setTickInterval } from "../../../libs/scheduling.js";
 import { queueUnban } from "../../../commands/moderation/unban.js";
 
 const World = world;
