+++ scripts/penrose/tickevent/clearlag/clearlag.ts
@@ -1,7 +1,7 @@
 import config from "../../../data/config.js";
 import { world, EntityQueryOptions, ItemStack, EntityItemComponent } from "mojang-minecraft";
 import { sendMsg } from "../../../util.js";
-import { clearTickInterval, setTickInterval } from "../../../misc/scheduling.js";
+import { clearTickInterval, setTickInterval } from "../../../libs/scheduling.js";
 import { clearItems } from "../../../data/clearlag.js";
 
 const World = world;
