+++ scripts/penrose/tickevent/fly/fly_a.ts
@@ -1,6 +1,6 @@
 import { world, EntityQueryOptions, Location, BlockLocation, Block } from "mojang-minecraft";
 import { getScore, flag, crypto } from "../../../util.js";
-import { clearTickInterval, setTickInterval } from "../../../misc/scheduling.js";
+import { clearTickInterval, setTickInterval } from "../../../libs/scheduling.js";
 import config from "../../../data/config.js";
 
 const World = world;
