+++ scripts/penrose/tickevent/knockback/antikb_a.ts
@@ -1,7 +1,7 @@
 import { EntityInventoryComponent, world } from "mojang-minecraft";
 import { flag, crypto } from "../../../util.js";
 import config from "../../../data/config.js";
-import { clearTickInterval, setTickInterval } from "../../../misc/scheduling.js";
+import { clearTickInterval, setTickInterval } from "../../../libs/scheduling.js";
 
 const World = world;
